+++ R/check_process.R
@@ -77,8 +77,11 @@ check_process <- R6::R6Class(
       # In some cases, check subprocess might suffer from a race condition, when
       # process itself finished, but the final results of the last subcheck
       # are not yet available to parse. Therefore we allow the process to
-      # finalize only if the last subcheck has reported status.
-      if (checks[length(checks)] != "") {
+      # finalize only if the last subcheck has reported status. However
+      # if we stuck in this state for longer than 3 minutes we should
+      # try to finish anyway, to prevent possible infinite loops.
+      if (checks[length(checks)] != "" |
+          Sys.time() - private$time_finish >= as.difftime(3, units = "mins")) {
         self$save_results()
         private$cache_parsed_results()
         private$free_file_descriptors()